New release announcement style "Westin" by shadowm and vultraz
This is a work-in-progress modern announcement style intended to replace
the dated Glamdrol style. It is not quite finished (the large intro
background is a headache due to its sheer size and alignment issues),
but it's expected to be complete in a few days.

Some of the most important changes:

 * There is no TOC anymore. (vultraz)
 * There is a large intro section with a background picture and words
   and a text-less shield logo (vultraz/shadowm).
 * Big image button with a link to the downloads section. (vultraz)
 * Animated scroll from download button. (shadowm)
 * Gradients, gradients everywhere. (shadowm)
 * Some styles are overridden on small display devices. (shadowm)
 * There is a copyright note in the footer. (shadowm)
 * Added dependency on jQuery, might be dropped if I don't find more
   uses for jQuery. (shadowm)