Import styles and drop dependency on Glamdrol
We were only using something like 10% of the Glamdrol stylesheet, the
rest of it being mostly wiki-specific styles and other things we do not
use here (front page overview box, etc.). Since Glamdrol is so large, it
was pretty much a waste of bandwidth. And for people who normally visit
wesnoth.org and have Glamdrol cached, it still posed a problem:

 * Westin is *huge*. It @imports Glamdrol so as to block loading the
   rest of the stylesheet since things would otherwise look horribly
   broken if Glamdrol was downloaded and applied first.
 * Conversely, if Westin was loaded before Glamdrol (by <link>ing
   Glamdrol), things would look terribly broken.

Included in this commit is the Glamdrol/Monobook bullet, which is also
included inline in the stylesheet for browsers that support Base64
inline images.

Westin before and after this change is unchanged to the pixel level as
far as I an tell.