Rearrange header bg, apply a fading border, and make section divider thinner
Rearranging the background image was necessary to make it less
conspicuously misaligned relative to the logo. Now the logo isn't
aligned to anything, but it doesn't look like it's *supposed* to be
aligned either, which is the superior option.

The image is also now rearranged on low resolutions so that there is
about as much sky visible as in the high resolution ruleset.

By applying a fading border (and the background color from the divider),
it's less conspicuous when the image gets clipped at the bottom, and the
divider can be made thinner.