gettext.w.o: Config and shared functions clean-up/refactoring/documentation
Among other things, got rid of the silly implode/explode dance going on
with the package/textdomain lists that is really a waste of time (and
also made the source values poor diff material) since everything needs
them in array form in the end.

Also completely eliminated the issue of code duplication in config.php
when enumerating add-ons.