start/1.14: Initial template for the press release announcement
This is based on the 1.12 announcement, using placeholder text for the
most part, as well as replacing references to 1.12 with 1.14 and adding
a historical link to the 1.12 release announcement.

The stylesheet in use at the moment is Westin, but it's slated to be
replaced with Wesmere with a custom supplement on top to add a few
styles not needed elsewhere.