gettext.w.o: Use partial regex matches on msgfmt's diagnostics to gather stats
The full-line regex match approach has an issue where it doesn't cover
all possible cases (in particular, a catalogue that is all fuzzies or
one that is only fuzzies and untranslated strings), resulting in invalid
stats.

More importantly, hrubymar10 on Discord reported an issue with the stats
for certain languages displaying with an untranslated string count of 0
when a quick glance at the .po files reveals that to be untrue. This
turns out to have been introduced by a fix for PHP warnings from 7 years
ago (commit 662f30252fd6854a7a012249738bd62c11ed80d1) where the author
didn't accurately identify the expected layout of the match captures
array (which is perfectly understandable given that the gettext.w.o
codebase was unreadable shit until my refactoring last year).

Somehow I'm still undecided whether commit
501645fed9cd3e60e74c716519602e40b2345b3b was a good idea or whether it
even makes a difference, but I guess I'll let translators be the judge
for this one. Hopefully this time it won't take 7 years for them to
report back.

Thanks sinda for the simpler match algorithm idea and hrubymar10 for the
original issue report.