Revert "removed Old English, Asturian and Friulian from g.w.o"
I misunderstood how this list was used. The set of which languages
to show is determined by the languages provided by the relevant
Wesnoth build; this list is just to map the language codes to names.

The preceding commit that adds Ancient Greek is correct, that name
will be shown when the Wesnoth build provides "grc".

This reverts commit 1e6eb3477642e6d94da4a38a1abf8fb945104175.